IBM_BE/chatbot_economic_terms/requirements.txt
langchain==0.2.6
ibm-watsonx-ai==1.0.10
langchain_ibm==0.1.8
langchain_community==0.2.6
sentence-transformers==3.0.1
chromadb==0.5.3
pydantic==2.8.2
langchain-huggingface==0.0.3
python-dotenv==1.0.1
pymupdf4llm==0.0.13
openpyxl==3.1.5
